pyproject.toml
[tool.poetry]
name = "travel"
version = "0.1.0"
description = ""
authors = ["Ismail Gadzhiev <ismailgadzhievff@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^4.8.0"
bs4 = "^0.0.1"
requests = "^2.28.2"
ipykernel = "^6.21.2"
pyyaml = "^6.0"
pandas = "^1.5.3"
apache-airflow = "^2.5.1"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "^23.1.0"
openpyxl = "^3.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
